+++ tuplities/src/lib.rs
@@ -16,11 +16,11 @@ pub mod prelude {
     #[cfg(feature = "flatten-nest")]
     pub use tuplities_flatten_nest::{
         FlattenMatrixElements, FlattenNestedTuple, FlattenNestedTupleMatrix, IntoNestedTupleOption,
-        NestMatrixElements, NestTuple, NestTupleMatrix, NestedTupleIndex, NestedTupleIndexMut,
-        NestedTupleMut, NestedTupleOption, NestedTupleOptionWith, NestedTuplePopBack,
-        NestedTuplePopFront, NestedTuplePushBack, NestedTuplePushFront, NestedTupleRef,
-        NestedTupleReplicate, NestedTupleRow, NestedTupleRowMut, NestedTupleTryFrom,
-        NestedTupleTryInto,
+        NestMatrixElements, NestTuple, NestTupleMatrix, NestedTupleFrom, NestedTupleIndex,
+        NestedTupleIndexMut, NestedTupleInto, NestedTupleMut, NestedTupleOption,
+        NestedTupleOptionWith, NestedTuplePopBack, NestedTuplePopFront, NestedTuplePushBack,
+        NestedTuplePushFront, NestedTupleRef, NestedTupleReplicate, NestedTupleRow,
+        NestedTupleRowMut, NestedTupleTryFrom, NestedTupleTryInto,
     };
     #[cfg(feature = "from")]
     pub use tuplities_from::{TupleFrom, TupleInto};